requirements.txt
langchain
langchain-ollama
streamlit
langchain-experimental
langchain-community
PyPDF2
python-docx